Simplify demo and fix packaging domains
Simplify demo by refactoring from recycler to basic player, fix packaging domains, and update README's
