+++ VUDRMWidevineDemo/app/build.gradle
@@ -1,41 +1,34 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 29
-
-    defaultConfig {
-        applicationId "com.jwplayer.demo.recyclerview"
-        minSdkVersion 19
-        targetSdkVersion 29
-        versionCode 1
-        versionName "1.0"
-        multiDexEnabled true
-    }
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
-    compileOptions {
-        sourceCompatibility = '1.8'
-        targetCompatibility = '1.8'
-    }
+	compileSdkVersion 29
+	
+	defaultConfig {
+		minSdkVersion 19
+		targetSdkVersion 29
+		versionCode 1
+		versionName "1.0"
+		applicationId "com.jwplayer.vudrmwidevine"
+	}
+	buildTypes {
+		release {
+			minifyEnabled false
+			proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+		}
+	}
+	compileOptions {
+		sourceCompatibility 1.8
+		targetCompatibility 1.8
+	}
 }
 
 dependencies {
-    implementation fileTree(include: ['*.jar'], dir: 'libs')
-    testImplementation 'junit:junit:4.13.2'
+	implementation 'com.android.support.constraint:constraint-layout:2.0.4'
 
-    implementation'androidx.appcompat:appcompat:1.2.0'
-    implementation'com.google.android.material:material:1.0.0'
-    implementation 'androidx.media:media:1.3.1'
-    implementation group: 'androidx.constraintlayout', name: 'constraintlayout', version: '2.0.4'
-    implementation group: 'androidx.recyclerview', name: 'recyclerview', version: '1.2.1'
-
-    implementation group: 'com.google.android.gms', name: 'play-services-auth', version: '18.1.0'
-
-    def sdkVersion = '3.19.3'
+	def sdkVersion = '3.19.3'
 	implementation "com.longtailvideo.jwplayer:jwplayer-core:${sdkVersion}"
-	implementation "com.longtailvideo.jwplayer:jwplayer-common:${sdkVersion}"
+	implementation ("com.longtailvideo.jwplayer:jwplayer-common:${sdkVersion}")
+	implementation('com.google.android.material:material:1.3.0') {
+		force = true
+	}
 }
